BREAK: completely refactor Plot, implement Figure
Separated responsibilities for clearer code and functionality based on
concepts of Figure/Plot/Axis/Curves etc. A figure may contain an
arbitrary number of plots (subplots), which may contain up to two axes
(main and twinx), which may contain an arbitrary number of lines. Figure
handles legend, and fonts.

BREAK: read() is renamed to load().
BREAK: files and labels must be specified in load() only. Not in Plot()
BREAK: Twin axis is now differently handled. A Plot() object with its own set
of curves can be specified (loaded), and given to another Plot object as
a twin. Drawing of the twin is handled by the main. This saves
unnecesary duplication of arguments to functions, and clarifies plotting
code.

See ./examples/s-tui for how to build plots with the updated interface.